+++ plutus-ledger-api/plutus-ledger-api.cabal
@@ -36,26 +36,26 @@ library
     hs-source-dirs: src
     default-language: Haskell2010
     exposed-modules:
-        Plutus.V1.Ledger.Address
-        Plutus.V1.Ledger.Api
-        Plutus.V1.Ledger.Bytes
-        Plutus.V1.Ledger.Contexts
-        Plutus.V1.Ledger.Credential
-        Plutus.V1.Ledger.Crypto
-        Plutus.V1.Ledger.DCert
-        Plutus.V1.Ledger.Interval
-        Plutus.V1.Ledger.Scripts
-        Plutus.V1.Ledger.Tx
-        Plutus.V1.Ledger.Time
-        Plutus.V1.Ledger.Value
-        Plutus.V1.Ledger.EvaluationContext
+        PlutusLedgerApi.V1
+        PlutusLedgerApi.V1.Address
+        PlutusLedgerApi.V1.Bytes
+        PlutusLedgerApi.V1.Contexts
+        PlutusLedgerApi.V1.Credential
+        PlutusLedgerApi.V1.Crypto
+        PlutusLedgerApi.V1.DCert
+        PlutusLedgerApi.V1.Interval
+        PlutusLedgerApi.V1.Scripts
+        PlutusLedgerApi.V1.Tx
+        PlutusLedgerApi.V1.Time
+        PlutusLedgerApi.V1.Value
+        PlutusLedgerApi.V1.EvaluationContext
 
-        Plutus.V2.Ledger.Api
-        Plutus.V2.Ledger.Contexts
-        Plutus.V2.Ledger.Tx
-        Plutus.V2.Ledger.EvaluationContext
+        PlutusLedgerApi.V2
+        PlutusLedgerApi.V2.Contexts
+        PlutusLedgerApi.V2.Tx
+        PlutusLedgerApi.V2.EvaluationContext
 
-        Plutus.ApiCommon
+        PlutusLedgerApi.Common
     other-modules:
         Codec.CBOR.Extras
         Data.Either.Extras
@@ -85,10 +85,10 @@ library plutus-ledger-api-testlib
     visibility: public
     hs-source-dirs: testlib
     exposed-modules:
-        Plutus.Ledger.Test.Examples
-        Plutus.Ledger.Test.EvaluationContext
-        Plutus.Ledger.Test.ProtocolVersions
-        Plutus.Ledger.Test.Scripts
+        PlutusLedgerApi.Test.Examples
+        PlutusLedgerApi.Test.EvaluationContext
+        PlutusLedgerApi.Test.ProtocolVersions
+        PlutusLedgerApi.Test.Scripts
     build-depends:
         base >=4.9 && <5,
         plutus-core ^>=1.0,
