+++ plutus-ledger-api/src/PlutusLedgerApi/V1.hs
@@ -1,7 +1,7 @@
 {- |
 The interface to Plutus V1 for the ledger.
 -}
-module Plutus.V1.Ledger.Api (
+module PlutusLedgerApi.V1 (
     -- * Scripts
     SerializedScript
     , Script
@@ -108,22 +108,23 @@ module Plutus.V1.Ledger.Api (
 ) where
 
 import Data.SatInt
-import Plutus.ApiCommon as Common hiding (evaluateScriptCounting, evaluateScriptRestricting, isScriptWellFormed)
-import Plutus.ApiCommon qualified as Common (evaluateScriptCounting, evaluateScriptRestricting, isScriptWellFormed)
-import Plutus.V1.Ledger.Address
-import Plutus.V1.Ledger.Bytes
-import Plutus.V1.Ledger.Contexts
-import Plutus.V1.Ledger.Credential
-import Plutus.V1.Ledger.Crypto
-import Plutus.V1.Ledger.DCert
-import Plutus.V1.Ledger.EvaluationContext
-import Plutus.V1.Ledger.Interval hiding (singleton)
-import Plutus.V1.Ledger.Scripts as Scripts
-import Plutus.V1.Ledger.Time
-import Plutus.V1.Ledger.Value
 import PlutusCore.Data qualified as PLC
 import PlutusCore.Evaluation.Machine.ExBudget as PLC
 import PlutusCore.Evaluation.Machine.ExMemory (ExCPU (..), ExMemory (..))
+import PlutusLedgerApi.Common as Common hiding (evaluateScriptCounting, evaluateScriptRestricting, isScriptWellFormed)
+import PlutusLedgerApi.Common qualified as Common (evaluateScriptCounting, evaluateScriptRestricting,
+                                                   isScriptWellFormed)
+import PlutusLedgerApi.V1.Address
+import PlutusLedgerApi.V1.Bytes
+import PlutusLedgerApi.V1.Contexts
+import PlutusLedgerApi.V1.Credential
+import PlutusLedgerApi.V1.Crypto
+import PlutusLedgerApi.V1.DCert
+import PlutusLedgerApi.V1.EvaluationContext
+import PlutusLedgerApi.V1.Interval hiding (singleton)
+import PlutusLedgerApi.V1.Scripts as Scripts
+import PlutusLedgerApi.V1.Time
+import PlutusLedgerApi.V1.Value
 import PlutusTx (FromData (..), ToData (..), UnsafeFromData (..), fromData, toData)
 import PlutusTx.Builtins.Internal (BuiltinData (..), builtinDataToData, dataToBuiltinData)
 import PlutusTx.Prelude (BuiltinByteString, fromBuiltin, toBuiltin)
