+++ plutus-ledger-api/test/Spec.hs
@@ -1,5 +1,13 @@
 module Main(main) where
 
+
+import PlutusLedgerApi.Test.EvaluationContext (evalCtxForTesting)
+import PlutusLedgerApi.Test.Examples
+import PlutusLedgerApi.V1
+import Spec.Builtins qualified
+import Spec.Eval qualified
+import Spec.Interval qualified
+
 import Test.Tasty
 import Test.Tasty.HUnit
 import Test.Tasty.QuickCheck
@@ -7,14 +15,8 @@ import Test.Tasty.QuickCheck
 import Control.Monad (void)
 import Data.Either
 import Data.Word (Word8)
-import Plutus.Ledger.Test.EvaluationContext (evalCtxForTesting)
-import Plutus.Ledger.Test.Examples
-import Plutus.Ledger.Test.ProtocolVersions
-import Plutus.V1.Ledger.Api
-import Spec.Builtins qualified
+import PlutusLedgerApi.Test.ProtocolVersions
 import Spec.CostModelParams qualified
-import Spec.Eval qualified
-import Spec.Interval qualified
 
 main :: IO ()
 main = defaultMain tests
